+++ src/Payloads/EditOrganizationNotificationSettingsPayload.php
@@ -1,0 +1,6 @@
+<?php
+
+namespace Shellrent\VeeamVspcApiClient\Payloads;
+
+class EditOrganizationNotificationSettingsPayload extends JsonPatchPayload {
+}