+++ src/Payloads/EditOrganizationPolicySettingsPayload.php
@@ -1,0 +1,6 @@
+<?php
+
+namespace Shellrent\VeeamVspcApiClient\Payloads;
+
+class EditOrganizationPolicySettingsPayload extends JsonPatchPayload {
+}